feat(workflow): add interact_terminal_session + PTY to durable agent
Mirrors the AI-SDK tool surface in the workflow durable agent: enables
`run_terminal_cmd interactive=true` and the `interact_terminal_session`
tool (view/wait/send/kill), and removes the workflow-only
`start_command_async`/`wait_command` pair. Action handlers and the
PTY-creation branch are extracted into shared impl modules so both code
paths call the same code. Workflow PTY skips live `data-terminal`
streaming — the sidebar reads `sessionSnapshot` from the final tool
result. PTY sessions are torn down in `agentRunWorkflow`'s finally block
to mirror chat-handler.onFinish lifecycle.
