+++ app/components/MessagePartHandler.tsx
@@ -189,10 +189,8 @@ export const MessagePartHandler = memo(function MessagePartHandler({
     case "tool-shell":
     case "tool-run_terminal_cmd":
     case "tool-interact_terminal_session":
-    // Long-running workflow agent (agent-long mode) shell tools
-    case "tool-run_command":
-    case "tool-start_command_async":
-    case "tool-wait_command": {
+    // Long-running workflow agent (agent-long mode) shell tool
+    case "tool-run_command": {
       const effectiveToolCallId =
         (part as any).data?.toolCallId ?? part.toolCallId;
       const precomputedStreamingOutput = effectiveToolCallId
