+++ lib/ai/tools/schemas.ts
@@ -78,11 +78,81 @@ export const RUN_TERMINAL_CMD_INPUT_SCHEMA = z.object({
     ),
 });
 
-// Workflow variant — durable steps cannot host an in-memory PTY session, so
-// the `interactive` field is omitted from the schema entirely.
-export const RUN_TERMINAL_CMD_WORKFLOW_INPUT_SCHEMA = z.object(
-  runTerminalCmdBaseShape,
-);
+// Workflow variant — same surface as the AI-SDK schema. The durable agent
+// uses the same in-memory `ptySessionManager` singleton as the chat handler;
+// PTY sessions live for one agent turn (closed in `agent-run.ts` finally).
+export const RUN_TERMINAL_CMD_WORKFLOW_INPUT_SCHEMA =
+  RUN_TERMINAL_CMD_INPUT_SCHEMA;
+
+// ── interact_terminal_session ───────────────────────────────────────────
+
+export const INTERACT_TERMINAL_SESSION_DEFAULT_WAIT_TIMEOUT_SECONDS = 10;
+export const INTERACT_TERMINAL_SESSION_MAX_WAIT_TIMEOUT_SECONDS = 300;
+
+export const INTERACT_TERMINAL_SESSION_DESCRIPTION = `Interact with persistent shell sessions in the sandbox environment.
+
+<supported_actions>
+- \`view\`: View the content of a shell session
+- \`wait\`: Wait for the running process in a shell session to return
+- \`send\`: Send input to the active process (stdin) in a shell session
+- \`kill\`: Terminate the running process in a shell session
+</supported_actions>
+
+<instructions>
+- Sessions are created by \`run_terminal_cmd\` with \`interactive=true\`; pass the returned \`session\` id here
+- When using \`view\` action, ensure command has completed execution before using its output
+- Set a short \`timeout\` (such as 5s) on \`wait\` for processes that don't return promptly to avoid meaningless waiting time
+- Processes are NEVER killed on timeout — they keep running in the session; \`timeout\` only controls how long to wait for output before returning
+- Use \`wait\` action when a process needs additional time to complete and return
+- Only use \`wait\` after \`send\` (or after \`run_terminal_cmd\` returned without finishing); decide whether to wait based on the prior output
+- DO NOT use \`wait\` for long-running daemon processes
+- \`send\` writes input and captures only the immediate response chunk; if the process needs more time before it replies, follow up with \`action=wait\`
+- \`input\` is sent verbatim. Without a trailing \\n (or \`Enter\`), the line is typed but NOT submitted — a follow-up \`send\` will append to the same line. ALWAYS include \\n unless you specifically want to type without pressing Enter (e.g. building up a key sequence)
+- For special keys, use official tmux key names: C-c (Ctrl+C), C-d (Ctrl+D), C-z (Ctrl+Z), Up, Down, Left, Right, Home, End, Escape, Tab, Enter, Space, F1-F12, PageUp, PageDown
+- For modifier combinations: M-key (Alt), C-S-key (Ctrl+Shift)
+- Note: Use official tmux names (BSpace not Backspace, DC not Delete, Escape not Esc)
+- For non-key strings in \`input\`, DO NOT perform any escaping; send the raw string directly
+- Raw input BYPASSES command guardrails; never forward untrusted content
+</instructions>
+
+<recommended_usage>
+- Use \`view\` to check shell session history and latest status
+- Use \`wait\` to wait for the completion of long-running commands
+- Use \`send\` to interact with processes that require user input (e.g., responding to prompts)
+- Use \`send\` with special keys like C-c to interrupt, C-d to send EOF
+- Use \`kill\` to stop background processes that are no longer needed
+- Use \`kill\` to clean up dead or unresponsive processes
+</recommended_usage>`;
+
+export const INTERACT_TERMINAL_SESSION_INPUT_SCHEMA = z.object({
+  action: z
+    .enum(["view", "wait", "send", "kill"])
+    .describe("The action to perform"),
+  brief: z
+    .string()
+    .describe(
+      "A one-sentence preamble describing the purpose of this operation",
+    ),
+  input: z
+    .string()
+    .optional()
+    .describe(
+      'Input text to send to the interactive session. Required for `send`. Sent verbatim — without a trailing \\n (or `Enter`) the line is typed but NOT submitted, and a subsequent `send` will append to the same line. To submit just Enter, pass `"Enter"` or `"\\n"`.',
+    ),
+  session: z
+    .string()
+    .describe(
+      "The unique identifier of the target shell session (returned by `run_terminal_cmd` with `interactive=true`)",
+    ),
+  timeout: z
+    .number()
+    .int()
+    .optional()
+    .default(INTERACT_TERMINAL_SESSION_DEFAULT_WAIT_TIMEOUT_SECONDS)
+    .describe(
+      `Timeout in seconds to wait for output. Only used for \`wait\` action. Defaults to ${INTERACT_TERMINAL_SESSION_DEFAULT_WAIT_TIMEOUT_SECONDS} seconds. Max ${INTERACT_TERMINAL_SESSION_MAX_WAIT_TIMEOUT_SECONDS} seconds.`,
+    ),
+});
 
 // ── file ────────────────────────────────────────────────────────────────
 
@@ -563,33 +633,4 @@ export const OPEN_URL_INPUT_SCHEMA = z.object({
     .describe(
       "A one-sentence preamble describing the purpose of this operation",
     ),
-});
-
-// ── start_command_async / wait_command (workflow-only) ──────────────────
-//
-// These two tools are exclusive to the durable workflow agent. The normal
-// AI-SDK agent uses `run_terminal_cmd` with `is_background: true` instead.
-// Kept here for symmetry with the rest of the schema/description module.
-
-export const START_COMMAND_ASYNC_DESCRIPTION = `Start a long-running shell command in the background. REQUIRED args: command (string), outputFile (absolute path string). Returns { result: { handle, outputFile } } — pass the handle to wait_command. Use this for any scan that may exceed 60 seconds (nmap -A, sqlmap, ffuf, gobuster, nikto, hydra, etc.).`;
-
-export const START_COMMAND_ASYNC_INPUT_SCHEMA = z.object({
-  command: z
-    .string()
-    .describe("Shell command to execute, e.g. 'nmap -A example.com'."),
-  outputFile: z
-    .string()
-    .describe(
-      "Absolute path inside the sandbox where stdout+stderr will be redirected, e.g. '/home/user/scan.log'.",
-    ),
-});
-
-export const WAIT_COMMAND_DESCRIPTION = `Wait for a backgrounded command (started via start_command_async) to finish. Polls every interval seconds, up to maxMinutes.`;
-
-export const WAIT_COMMAND_INPUT_SCHEMA = z.object({
-  handle: z.string(),
-  outputFile: z.string(),
-  intervalSeconds: z.number().int().min(5).max(120).default(30),
-  maxMinutes: z.number().int().min(1).max(120).default(30),
-  tailLines: z.number().int().min(10).max(2000).default(200),
 });
